Added files via upload
Adding an optional "datetimefunc" option, an alternative to "datetime", allowing the user to specify a function to call to return a Date object, rather than declaring a single Date object directly.
The current file (or list of files to be concatenated) is passed in as an array, as the first argument.  A Date object is expected in return.